feat: add support for translation feature in Zalo PC/Web with ZaDark extension
The commit adds support for a new feature called "Dịch tin nhắn" (Translation) in Zalo PC/Web. This feature allows users to translate messages in Zalo to improve communication.

The changes include:
- Added a new file "dich-tin-nhan-zalo.mdx" in the "blog" directory to provide a tutorial on how to use the translation feature.
- Updated the existing file "cach-tat-dang-soan-tin-tren-zalo/index.mdx" to improve the content and formatting.

The commit also includes some minor changes to improve the readability and user experience of the blog posts.

Note: The commit message assumes that the changes in the "dich-tin-nhan-zalo.mdx" file are new and not a modification of an existing file.